W250: the lever checklist in unswept files — +2 MATCH, and the near-MATCH band is now genuinely exhausted
**6602 -> 6604 first-party MATCH.** 0 lost, 0 new over-window, 0 regressions.

Wave 249 established that an identical prologue is not proof of a compiler floor, and that an
eight-lever source-shape checklist recovers matches the prologue test throws away. This wave applied
that checklist to the ~46 files no previous sweep had touched.

**Two more MATCHes, and neither came from the lever I predicted:**

- **`h_sfdply 0010e880` nd27 -> 0**, purely by **local declaration order** — reordering the pointer and
  scalar declarations to match retail's live-value colouring. 976/976.
- **`mt_scene 003b5980` nd73 -> 0**, from using a call result directly rather than through a retained
  result local, combined with narrowing and type corrections.

I had told agents to lead with levers 6 and 7 (explicit narrowing, declared type) because those
produced both of wave 249's matches. **They produced neither of this wave's.** Across four waves the
scoreboard is now: declaration order 2 matches, narrowing/type 2, early-return guard 1, result-local
plus structured break 1, assignment-in-condition 1 (shared). **No lever dominates — the checklist works
because it is exhaustive, not because any entry is privileged.** That is the durable lesson.

Other real gains: **`k_unit 001ce880` nd132 -> 14** (restoring an explicit `hasResources` result and
normalization plus a typed `Model` local; 200 -> 220 of a 224 window); `k_fldFrame 001ae580` nd3622 ->
3555 and 60 bytes smaller, by correcting `(f32)(u32)work->turnMode` to `(f32)work->turnMode` — the
field is declared `s32`, so the unsigned cast was inventing a bias expansion; `h_maestro 00112110`
nd13 -> 2; `fclCombineMisc 003d6ae0` nd18 -> 11 by copying declaration order from the adjacent matched
aggregate-copy loop.

## The band is now exhausted, and that is worth stating plainly

Around fifty functions were audited through all eight levers this wave with per-lever measurements
recorded, on top of forty last wave. The floor verdicts are now backed by evidence rather than by a
single prologue check, and the residuals cluster tightly into the same measured families:
call-argument setup order before a `JAL` (by far the most common), cyclic saved-register colouring
with matching frame and save count, loop-preheader constant placement, commutative `addu`/`mul.s`
orientation, `addiu` versus `daddiu`, two adjacent independent loads reversed, `slti $at` versus `$v1`,
FPR bank rotation, and branch booleanization.

Many probes actively regressed and were reverted, which is itself evidence the band is worked out: a
typed 4-argument alias took `fclCombineEffect 0041a730` from nd51 to 176; a switch conversion took
`0041ae20` to nd119; direct typed `f32` indexing took `mt_evtLoadSave 0038e660` from nd13 to 179;
replacing five materialized absolute-pointer locals with per-site loads took `mt_evtEffect 00395570`
from nd54 to 282; a declaration-order move took `btlTarget 002db480` from nd7 to 15.

**`_h_camp_main 00136750` is now definitively closed.** Its only remaining difference is the 0x40
versus 0x50 frame allocation word. Wave 247 proved retail has no load, store or address formation
anywhere in the extra 0x10 and saves only s0/s1/s2 and ra; this wave confirmed no source-shape lever
reaches it. Growing the frame would require an unused local, which is banned. It stays unmatched.

+++ src/Kosaka/Field/k_fldFrame.c
@@ -1214,7 +1214,7 @@ void* func_001ae580(KwlnTask* task)
                 localAngleB = amount;
             }
             work->frame = 0;
-            work->angleStep = localAngleB / (f32)(u32)work->turnMode;
+            work->angleStep = localAngleB / (f32)work->turnMode;
             work->currentAngle = localAngle;
         }
         else
